Slow heartbeat for hw wallet and generate hwdev.txt file
The heartbeat was overloading the rpc wallet with calls every 5 seconds
which resulted in an ever growing queue of messages. Have slowed the
hearbeat down significantly for hardware wallets.

Have also added code to ensure that the hwdev.txt file gets created upon
wallet creation, the GUI wallet needs this to know if a wallet is a
hardware wallet and this isn't generated by default unless we pass a
name to the initial RPC create wallet call.