Update algorithm counting contributors
The number of contributions to a service node is dependant on the number
of locked contributions instead of contributors. The difference being
that a single contributor may make several locked contributions to a
service node which will reduce the number of available slots. To clarify
a single contributor may contribute twice to a service node, using 2
slots rather than one.